+++ src/main/kotlin/de/triplet/gradle/play/ImageFileFilter.kt
@@ -1,0 +1,8 @@
+package de.triplet.gradle.play
+
+import java.io.File
+import java.io.FileFilter
+
+internal class ImageFileFilter : FileFilter {
+    override fun accept(file: File) = file.extension.toLowerCase() in IMAGE_EXTENSIONS
+}