+++ src/main/kotlin/de/triplet/gradle/play/PlayAccountConfig.kt
@@ -1,0 +1,11 @@
+package de.triplet.gradle.play
+
+import java.io.File
+
+open class PlayAccountConfig(var name: String = "") {
+    var serviceAccountEmail: String? = null
+
+    var pk12File: File? = null
+
+    var jsonFile: File? = null
+}