+++ src/main/kotlin/de/triplet/gradle/play/PlayPublisherPlugin.kt
@@ -1,77 +1,75 @@
 package de.triplet.gradle.play
 
 import com.android.build.gradle.AppExtension
-import com.android.build.gradle.AppPlugin
+import com.android.builder.model.BaseConfig
+import groovy.lang.GroovyObject
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
-class PlayPublisherPlugin implements Plugin<Project> {
+class PlayPublisherPlugin : Plugin<Project> {
+    override fun apply(project: Project) {
+        val log = project.logger
+        val android = project.extensions.getByType(AppExtension::class.java)
+                ?: throw IllegalStateException("The 'com.android.application' plugin is required.")
+        val extension = project.extensions.create("play", PlayPublisherPluginExtension::class.java)
 
-    static final PLAY_STORE_GROUP = 'Play Store'
-    static final RESOURCES_OUTPUT_PATH = 'build/outputs/play'
+        android.extensions.add("playAccountConfigs", project.container(PlayAccountConfig::class.java))
 
-    @Override
-    void apply(Project project) {
-        def log = project.logger
-
-        if (!project.plugins.any { p -> p instanceof AppPlugin }) {
-            throw new IllegalStateException('The \'com.android.application\' plugin is required.')
-        }
-
-        def android = project.extensions.getByType(AppExtension)
-        def extension = project.extensions.create('play', PlayPublisherPluginExtension)
-
-        android.extensions.playAccountConfigs = project.container(PlayAccountConfig)
-
-        android.defaultConfig.ext.playAccountConfig = null
+        android.defaultConfig.extras.set("playAccountConfig", null)
 
         android.productFlavors.whenObjectAdded { flavor ->
-            flavor.ext.playAccountConfig = android.defaultConfig.ext.playAccountConfig
+            //flavor.ext.playAccountConfig = android.defaultConfig.ext.playAccountConfig
+            flavor.extras.set("playAccountConfig", android.defaultConfig.extras.get("playAccountConfig"))
         }
 
         android.applicationVariants.whenObjectAdded { variant ->
-            if (variant.buildType.isDebuggable()) {
+            if (variant.buildType.isDebuggable) {
                 log.debug("Skipping debuggable build type ${variant.buildType.name}.")
-                return
+                return@whenObjectAdded
             }
 
-            def flavorAccountConfig = variant.productFlavors.find { it.playAccountConfig }?.playAccountConfig
-            def defaultAccountConfig = android.defaultConfig.ext.playAccountConfig
-            def playAccountConfig = flavorAccountConfig ?: defaultAccountConfig
+            val flavorAccountConfig = variant.productFlavors
+                    .map { it as BaseConfig as GroovyObject }
+                    .mapNotNull { it.getProperty("playAccountConfig") }
+                    .firstOrNull() as? PlayAccountConfig
+            val defaultAccountConfig = android.defaultConfig.extras.get("playAccountConfig") as? PlayAccountConfig
+            val playAccountConfig = flavorAccountConfig ?: defaultAccountConfig ?: PlayAccountConfig()
 
-            def bootstrapTaskName = "bootstrap${variant.name.capitalize()}PlayResources"
-            def playResourcesTaskName = "generate${variant.name.capitalize()}PlayResources"
-            def publishApkTaskName = "publishApk${variant.name.capitalize()}"
-            def publishListingTaskName = "publishListing${variant.name.capitalize()}"
-            def publishTaskName = "publish${variant.name.capitalize()}"
+            val bootstrapTaskName = "bootstrap${variant.name.capitalize()}PlayResources"
+            val playResourcesTaskName = "generate${variant.name.capitalize()}PlayResources"
+            val publishApkTaskName = "publishApk${variant.name.capitalize()}"
+            val publishListingTaskName = "publishListing${variant.name.capitalize()}"
+            val publishTaskName = "publish${variant.name.capitalize()}"
 
             // Create and configure bootstrap task for this variant.
-            def bootstrapTask = project.tasks.create(bootstrapTaskName, BootstrapTask)
+            val bootstrapTask = project.tasks.create(bootstrapTaskName, BootstrapTask::class.java)
             bootstrapTask.extension = extension
             bootstrapTask.variant = variant
             if (!variant.flavorName.isEmpty()) {
                 bootstrapTask.outputFolder = project.file("src/${variant.flavorName}/play")
             } else {
-                bootstrapTask.outputFolder = project.file('src/main/play')
+                bootstrapTask.outputFolder = project.file("src/main/play")
             }
             bootstrapTask.playAccountConfig = playAccountConfig
             bootstrapTask.description = "Downloads the play store listing for the ${variant.name.capitalize()} build. No download of image resources. See #18."
             bootstrapTask.group = PLAY_STORE_GROUP
 
             // Create and configure task to collect the play store resources.
-            def playResourcesTask = project.tasks.create(playResourcesTaskName, GeneratePlayResourcesTask)
-            playResourcesTask.inputs.file(project.file('src/main/play'))
+            val playResourcesTask = project.tasks.create(playResourcesTaskName, GeneratePlayResourcesTask::class.java)
+
+            playResourcesTask.inputs.file(project.file("src/main/play"))
             if (!variant.flavorName.isEmpty()) {
                 playResourcesTask.inputs.file(project.file("src/${variant.flavorName}/play"))
             }
             playResourcesTask.inputs.file(project.file("src/${variant.buildType.name}/play"))
             playResourcesTask.inputs.file(project.file("src/${variant.name}/play"))
-            playResourcesTask.outputFolder = project.file("${RESOURCES_OUTPUT_PATH}/${variant.name}")
+
+            playResourcesTask.outputFolder = project.file("$RESOURCES_OUTPUT_PATH/${variant.name}")
             playResourcesTask.description = "Collects play store resources for the ${variant.name.capitalize()} build"
             playResourcesTask.group = PLAY_STORE_GROUP
 
             // Create and configure publisher meta task for this variant
-            def publishListingTask = project.tasks.create(publishListingTaskName, PlayPublishListingTask)
+            val publishListingTask = project.tasks.create(publishListingTaskName, PlayPublishListingTask::class.java)
             publishListingTask.extension = extension
             publishListingTask.playAccountConfig = playAccountConfig
             publishListingTask.variant = variant
@@ -80,11 +78,11 @@ class PlayPublisherPlugin implements Plugin<Project> {
             publishListingTask.group = PLAY_STORE_GROUP
 
             // Attach tasks to task graph.
-            publishListingTask.dependsOn playResourcesTask
+            publishListingTask.dependsOn(playResourcesTask)
 
-            if (variant.isSigningReady()) {
+            if (variant.isSigningReady) {
                 // Create and configure publisher apk task for this variant.
-                def publishApkTask = project.tasks.create(publishApkTaskName, PlayPublishApkTask)
+                val publishApkTask = project.tasks.create(publishApkTaskName, PlayPublishApkTask::class.java)
                 publishApkTask.extension = extension
                 publishApkTask.playAccountConfig = playAccountConfig
                 publishApkTask.variant = variant
@@ -92,19 +90,18 @@ class PlayPublisherPlugin implements Plugin<Project> {
                 publishApkTask.description = "Uploads the APK for the ${variant.name.capitalize()} build"
                 publishApkTask.group = PLAY_STORE_GROUP
 
-                def publishTask = project.tasks.create(publishTaskName)
+                val publishTask = project.tasks.create(publishTaskName)
                 publishTask.description = "Updates APK and play store listing for the ${variant.name.capitalize()} build"
                 publishTask.group = PLAY_STORE_GROUP
 
                 // Attach tasks to task graph.
-                publishTask.dependsOn publishApkTask
-                publishTask.dependsOn publishListingTask
-                publishApkTask.dependsOn playResourcesTask
-                publishApkTask.dependsOn variant.assemble
+                publishTask.dependsOn(publishApkTask)
+                publishTask.dependsOn(publishListingTask)
+                publishApkTask.dependsOn(playResourcesTask)
+                publishApkTask.dependsOn(variant.assemble)
             } else {
                 log.warn("Signing not ready. Did you specify a signingConfig for the variation ${variant.name.capitalize()}?")
             }
         }
     }
-
 }